Move FigureClickmenu
